implement kinde auth

+++ features/tasks/actions/task-actions.ts
@@ -1,11 +1,27 @@
 "use server";
 
-import { prisma } from "@/prisma/db";
-import { Prisma } from "@prisma/client";
+// next imports
 import { revalidatePath } from "next/cache";
+import { redirect } from "next/navigation";
 
-// Maybe Zod implementation instead of FormData
+// prisma imports
+import { Prisma } from "@prisma/client";
+import { prisma } from "@/prisma/db";
+
+// kinde imports
+import { getKindeServerSession } from "@kinde-oss/kinde-auth-nextjs/server";
+
+// idea: Zod implementation instead of FormData
 export async function createTask(formData: FormData) {
+  // authentication check
+  const { isAuthenticated, getUser } = getKindeServerSession();
+  if (!(await isAuthenticated())) {
+    redirect("/api/auth/login");
+  }
+
+  // get user
+  const user = await getUser();
+
   try {
     await prisma.task.create({
       data: {
@@ -13,11 +29,7 @@ export async function createTask(formData: FormData) {
         slug: (formData.get("title") as string)
           .replace(/\s+/g, "-")
           .toLowerCase(),
-        user: {
-          connect: {
-            email: "vsdutraa@gmail.com",
-          },
-        },
+        userId: user.id,
       },
     });
   } catch (error) {
@@ -34,6 +46,25 @@ export async function createTask(formData: FormData) {
 
 // ID is cuid. That's why it's a string.
 export async function editTask(formData: FormData, id: string) {
+  // authentication check
+  const { isAuthenticated, getUser } = getKindeServerSession();
+  if (!(await isAuthenticated())) {
+    redirect("/api/auth/login");
+  }
+
+  // authorization check
+  const user = await getUser();
+  const task = await prisma.task.findUnique({
+    where: { id },
+  });
+  if (task?.userId !== user.id) {
+    throw new Error("You're not authorized to edit this task.");
+  }
+
+  if (!task) {
+    throw new Error("Task with specified ID does not exist.");
+  }
+
   await prisma.task.update({
     where: { id },
     data: {
@@ -42,7 +73,27 @@ export async function editTask(formData: FormData, id: string) {
   });
 }
 
+// ID is cuid. That's why it's a string.
 export async function deleteTask(id: string) {
+  // authentication check
+  const { isAuthenticated, getUser } = getKindeServerSession();
+  if (!(await isAuthenticated())) {
+    redirect("/api/auth/login");
+  }
+
+  // authorization check
+  const user = await getUser();
+  const task = await prisma.task.findUnique({
+    where: { id },
+  });
+  if (task?.userId !== user.id) {
+    throw new Error("You're not authorized to edit this task.");
+  }
+
+  if (!task) {
+    throw new Error("Task with specified ID does not exist.");
+  }
+
   try {
     await prisma.task.delete({ where: { id } });
   } catch (error) {
